refactor(rename): IndexController에서 FrontendUrlRequestHandler로 이름 변경
패키지도 이동 (global.utils.web)
